export observation valuations as signed integers

+++ payntbind/src/synthesis/pomdp_family/ObservationEvaluator.h
@@ -31,13 +31,13 @@ namespace synthesis {
         std::vector<uint32_t> state_to_obs_class;
 
         /** Get the value of the observation expression in the given observation class. */
-        uint32_t observationClassValue(uint32_t obs_class, uint32_t obs_expr);
+        int64_t observationClassValue(uint32_t obs_class, uint32_t obs_expr);
 
         /**
          * Collect an observation valuation to a vector. Boolean expressions are represented as 0/1.
          * @note this is an observation valuation placeholder
          */
-        std::vector<uint32_t> observationValuation(uint32_t obs_class);
+        std::vector<int64_t> observationValuation(uint32_t obs_class);
 
         /**
          * Create a sub-POMDP from the given sub-MDP by associating its states with observations.
